refactor: add Highlighted type
