+++ packages/autocomplete-plugin-recent-searches/src/createLocalStorageRecentSearchesPlugin.ts
@@ -5,7 +5,7 @@ import {
   CreateRecentSearchesPluginParams,
   RecentSearchesPluginData,
 } from './createRecentSearchesPlugin';
-import { RecentSearchesItem } from './types';
+import { Highlighted, RecentSearchesItem } from './types';
 import {
   LOCAL_STORAGE_KEY,
   createLocalStorage,
@@ -33,7 +33,7 @@ export type CreateRecentSearchesLocalStorageOptions<
   /**
    * Function to search in the recent items.
    */
-  search?(params: SearchParams<TItem>): TItem[];
+  search?(params: SearchParams<TItem>): Array<Highlighted<TItem>>;
 };
 
 type LocalStorageRecentSearchesPluginOptions<
