+++ packages/autocomplete-plugin-recent-searches/src/types/Highlighted.ts
@@ -1,0 +1,7 @@
+export type Highlighted<TItem> = TItem & {
+  _highlightResult: {
+    query: {
+      value: string;
+    };
+  };
+};
