--- packages/autocomplete-plugin-recent-searches/src/types/RecentSearchesHit.ts
@@ -1,9 +1,0 @@
-import { RecentSearchesItem } from './RecentSearchesItem';
-
-export type RecentSearchesHit = RecentSearchesItem & {
-  _highlightResult: {
-    query: {
-      value: string;
-    };
-  };
-};
